requirements.txt
requests
pandas
google-cloud-bigquery
pyarrow
pandas-gbq
